requirements.txt
Django==1.10.6
celery==4.0.2
dj-database-url==0.4.2
setuptools==35.0.2
django-suit==0.2.25
psycopg2==2.7.1
django-import-export==0.5.1
django-mptt==0.8.7
django-haystack==2.5.1
django-phonenumber-field==1.3.0pip
WeasyPrint==0.36
cffi-1.10.0
WeasyPrint==0.36
elasticsearch==5.0.1
whitenoise==3.3.0
gunicorn==19.7.1

